% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws.demc.r
\name{draws.demc}
\alias{draws.demc}
\title{Extracts the draws from a demc object}
\usage{
draws.demc(object, keep.all = 0.9, thin = 1)
}
\arguments{
\item{object}{\code{demc} object}

\item{keep.all}{logical or fraction of draws to keep. If fraction, the one-complement is discarded as burnin.
If logical the first half of the draws are discarded. Default 0.9.}

\item{thin}{integer (default 1), keep every thin-th draw}
}
\value{
a dataframe with parameters as variables and as rows the kept samples
}
\description{
Extracts the draws from a demc object
}
