% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.demc.r
\name{summary.demc}
\alias{summary.demc}
\title{Summarizes a demc object}
\usage{
\method{summary}{demc}(a, trans = NULL, keep.all = 0.9,
  Rupper.keep = FALSE)
}
\arguments{
\item{trans}{a vector of length k (default NULL):  "" if no transformation, or "log" or "logit"}

\item{keep.all}{logical or fraction of draws to keep. If fraction, the one-complement is discarded as burnin.
If logical the first half of the draws are discarded. Default 0.9.}

\item{Rupper.keep}{logical (default FALSE); if TRUE:  keep Rupper,  (Otherwise don't display it.)}

\item{object}{\code{demc} object}
}
\value{
an S3 object of class \code{demc} which is a list with 

$summary 
$accept.prob.mat
}
\description{
Summarizes a demc object
}
\details{
The summary is adapted from the monitor function written by Andrew Gelman.
}
